requirements.txt
aiofile==3.8.8
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
caio==0.9.17
certifi==2024.7.4
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-util==1.2.1
redis==5.0.8
six==1.16.0
typing_extensions==4.12.2
yarl==1.9.4
